补充 Widget_BackgroundTaskList 接入相关信息，并修改 ModFileApplication 完成后台任务的显示;

+++ L4d2_Mod_Manager_Tool/DTO/BackgroundTaskProgress.cs
@@ -10,7 +10,7 @@ namespace L4d2_Mod_Manager_Tool.DTO
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Progress { get; set; }
+        public int Progress { get; set; }
         public string Status {get;set;}
     }
 }
